requirements.txt
Django==4.1.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
ipython==8.5.0
Pillow==9.2.0
django_filter==22.1
